// src/lib/types/post.d.ts
export type Post = {
  AIsummary: string;
  author: string;
  authorUUID: string;
  content: string;
  coverImageFile: File | null;
  coverImageLink: string;
  coverImageLinkAvif: string;
  coverImageLinkWebp: string;
  coverImageLinkPng: string;
  coverImageLink100x300Avif: string;
  coverImageLink100x300Webp: string;
  coverImageLink100x300Png: string;
  coverImageLink200x600Avif: string;
  coverImageLink200x600Webp: string;
  coverImageLink200x600Png: string;
  coverImageAlt: string;
  date: string;
  minToRead: number;
  slug: string;
  subTitle: string;
  title: string;
  topic: string;
  topicColor: string;
  uid: string;
};